requirements.txt
# Django & DRF
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.21.0
python-decouple==3.8

# Database
mongoengine==0.29.1
pymongo==4.6.0

# Authentication
djangorestframework-simplejwt==5.3.2

# FastAPI (Modern async framework)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async & Background Tasks
celery==5.3.4
redis==5.0.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9

# Security & Validation
python-dotenv==1.0.0
PyJWT==2.8.1
cryptography==41.0.7

# API Documentation
drf-spectacular==0.26.5

# Monitoring & Logging
sentry-sdk==1.38.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
factory-boy==3.3.0

# Development
black==23.12.1
flake8==6.1.0
isort==5.13.2
pylint==3.0.3

# Optional: Blockchain Integration
web3==6.11.2

# Optional: Email
django-anymail==10.1

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0
